requirements.txt
streamlit==1.28.0
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.0
plotly==5.17.0
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
reportlab==4.0.4
fpdf2==2.7.6
Pillow==10.0.0
requests==2.31.0